requirements.txt
matplotlib==3.0.0
numpy==1.14.3
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
pandas==0.23.4
PyYAML==3.13
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
sklearn==0.0
sympy==1.3
tqdm==4.26.0